+++ src/services/database/block.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model, Types } from 'mongoose';
-import { Block, BlockDocument } from '../schemas/block.schema';
+import { Block, BlockDocument } from '../../schemas/block.schema';
 
 @Injectable()
 export class BlockService {
