+++ src/services/database/news.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model, Types } from 'mongoose';
-import { News, NewsDocument } from '../schemas/news.schema';
+import { News, NewsDocument } from '../../schemas/news.schema';
 
 @Injectable()
 export class NewsService {
